Move keywords out of NVD collector
